upgrade to 0.2.1
