+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "zellular"
-version = "0.2.0"
+version = "0.2.1"
 description = "SDK package for zellular sequencer"
 readme = "README.md"
 requires-python = ">=3.6"
