+++ tests/nodes.json
@@ -1,29 +1,57 @@
 {
-    "0x0f36cb075aee6ed5eacdb90a778f6ff929ba03fc": {
-        "id": "0x0f36cb075aee6ed5eacdb90a778f6ff929ba03fc",
-        "public_key_g2": "1 5843676833903350800258463566472133286765173966425078700723295454382955351237 20691278963634651578031563010214351732131677830832211548383861034784340981267 8144307982224982106094096267890006774203396141489170067881750052924747094152 21844207826321315508581933057392271388691384756250607048367019522834704527753",
-        "address": "0x0f36cb075aee6ed5eacdb90a778f6ff929ba03fc",
+    "0x0985aa64c1baebe643a8c9b273f520da75ffaa06": {
+        "id": "0x0985aa64c1baebe643a8c9b273f520da75ffaa06",
+        "pubkeyG2_X": [
+            15476857764318855250348861691385413449579762526280009118002471392365988397482,
+            4038296986528022637840850934283081733190643927601197326810043846026821876992
+        ],
+        "pubkeyG2_Y": [
+            16790126939825669687176065900941664153943303573008828372749779879833513370168,
+            19448986540722803323791398667133466426149138321527515647500106105279751751127
+        ],
+        "address": "0x0985aa64c1baebe643a8c9b273f520da75ffaa06",
         "socket": "http://localhost:6001",
         "stake": 10
     },
-    "0x1433743e0a68d7b665c2b647bce136f2f6e46350": {
-        "id": "0x1433743e0a68d7b665c2b647bce136f2f6e46350",
-        "public_key_g2": "1 16581017045641723807396265106200287824951249207934753861000287012476969293946 15418811225341053798594871575665035150371471587756635311903495929043740811767 10327216150662745223550284616768579545535352008851123261976185161238930848579 11605505391894596607310358509522502860362552022905152118063087569255509133858",
-        "address": "0x1433743e0a68d7b665c2b647bce136f2f6e46350",
+    "0x722fc5411f54d7b5686ed35093bbb527e0150677": {
+        "id": "0x722fc5411f54d7b5686ed35093bbb527e0150677",
+        "pubkeyG2_X": [
+            13908695953677148157437222245300425944450648651727157079534962437467264728666,
+            2839315594107535084588809939161874910299607779835614843443224680902650971469
+        ],
+        "pubkeyG2_Y": [
+            20709463269904342846102307497584533754639131111820474763740481252988240599626,
+            2438130303266171282532076273846159478573013608211386666373780632240813919438
+        ],
+        "address": "0x722fc5411f54d7b5686ed35093bbb527e0150677",
         "socket": "http://localhost:6002",
         "stake": 10
     },
-    "0x387f79218312a26c376249dcb33a057f80c9e0d8": {
-        "id": "0x387f79218312a26c376249dcb33a057f80c9e0d8",
-        "public_key_g2": "1 20763527447656649358013819023147829570827747897533039040455886199289438436927 17272423058437753118861159904219667187292616386241867081665122224803344042444 4267308689741391185248338892701429197017396327863149536431778859282270272021 10784737954335926400966236174565553274003718231394643392074634862809512746772",
-        "address": "0x387f79218312a26c376249dcb33a057f80c9e0d8",
+    "0xc9125e11bc1848586f831bcd1b62edeb71d66540": {
+        "id": "0xc9125e11bc1848586f831bcd1b62edeb71d66540",
+        "pubkeyG2_X": [
+            19891739138404785882678287233072940048457245997072147557223116364004811168748,
+            863968383549500633672703415919285120274940005694861520897169864961838089223
+        ],
+        "pubkeyG2_Y": [
+            11512275551697273504484213285961272473380219120686483277502701887683759030404,
+            322410733407591005657607051014446559677271006405275610809438887471133916270
+        ],
+        "address": "0xc9125e11bc1848586f831bcd1b62edeb71d66540",
         "socket": "http://localhost:6003",
         "stake": 10
     },
-    "0xbdb2e6b2ab864e63a88d75ebe8f2fadf440b75f1": {
-        "id": "0xbdb2e6b2ab864e63a88d75ebe8f2fadf440b75f1",
-        "public_key_g2": "1 4333245661120623865312562731909481521225732874179058502672075866102670464760 11316985814223291448589698221345368915473872583537495947292570291159304826404 2623199747967925902574543797389735322163931823125269475026237265380980407915 21001231620156017091630034136431538254247758798748307485670488869118323067461",
-        "address": "0xbdb2e6b2ab864e63a88d75ebe8f2fadf440b75f1",
+    "0xf2669e755aac2691796efa7320fcfd726d28f67a": {
+        "id": "0xf2669e755aac2691796efa7320fcfd726d28f67a",
+        "pubkeyG2_X": [
+            16794484888936412172560668978518602588920783958898879764039719549467709353592,
+            14074556044203183521579418015834006458193293591972593575850457531909668759672
+        ],
+        "pubkeyG2_Y": [
+            13116327993633056011840737974694154115601178449243537336638074105176453152247,
+            21847532510894389534034994361325552777892232269627980482842494721767169782888
+        ],
+        "address": "0xf2669e755aac2691796efa7320fcfd726d28f67a",
         "socket": "http://localhost:6004",
         "stake": 10
     }
