+++ zellular/networks/eigenlayer.py
@@ -97,12 +97,3 @@ class EigenlayerNetwork(Network):
     def _get_stake(cls, operator: dict[str, Any]) -> float:
         stake = int(operator.get("stake", 0)) / (10**18)
         return stake if operator.get("id") in cls.DEFAULT_NODES else min(stake, 1)
-
-    @staticmethod
-    def _get_g2_key(operator: dict[str, Any]) -> attestation.G2Point:
-        return attestation.G2Point(
-            operator["pubkeyG2_X"][0],
-            operator["pubkeyG2_X"][1],
-            operator["pubkeyG2_Y"][0],
-            operator["pubkeyG2_Y"][1],
-        )
